# docs/specs/api-specs/openapi-missing-endpoints.yaml
openapi: 3.0.3
info:
  title: Cortex API - Missing Endpoints
  version: 1.0.0
  description: OpenAPI specification for missing and incomplete API endpoints
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.cortex.example.com/api
    description: Production server
...
paths:
...
  /projects/{projectId}/agent-runs/{runId}:
    get:
      tags:
        - agents
      summary: Get agent run
      operationId: getAgentRun
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: Project ID
...
  schemas:
...
    AgentRun:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        workflowId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED, CANCELLED]
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: 'openapi-error-responses.yaml#/components/schemas/Error'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []